fix(migrator): register mssql driver inside migrator without circular imports
